refactor: split AI signatures into data and logic modules
Follow imperatives.py pattern — ai_signatures_data.py holds tool
definitions; ai_signatures.py holds detection API only.
